requirements.txt
attrs==18.1.0
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.5
Pillow==5.2.0
Werkzeug==0.14.1
uwsgi